Use symfony/panther only when is needed
Use symfony/panther only when the functional tests require JavaScript.
