+++ tests/Functional/BaseFunctionalTestCase.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\DoctrineMongoDBAdminBundle\Tests\Functional;
+
+use Sonata\DoctrineMongoDBAdminBundle\Tests\App\AppKernel;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+abstract class BaseFunctionalTestCase extends WebTestCase
+{
+    /**
+     * @var KernelBrowser
+     */
+    protected $client;
+
+    protected function setUp(): void
+    {
+        $this->client = static::createClient();
+        $this->client->followRedirects();
+    }
+
+    protected static function getKernelClass(): string
+    {
+        return AppKernel::class;
+    }
+}
